@@ -1,6 +1,6 @@
 {
   "name": "mkdocs-material",
-  "version": "7.3.6+insiders-3.2.3",
+  "version": "8.0.0+insiders-4.0.0",
   "description": "A Material Design theme for MkDocs",
   "keywords": [
     "mkdocs",
@@ -55,16 +55,16 @@
     "@types/fuzzaldrin-plus": "^0.6.2",
     "@types/html-minifier": "^4.0.1",
     "@types/lunr": "^2.3.4",
-    "@types/node": "^16.11.9",
+    "@types/node": "^16.11.10",
     "@types/resize-observer-browser": "^0.1.6",
-    "@types/sass": "^1.43.0",
+    "@types/sass": "^1.43.1",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
     "autoprefixer": "^10.4.0",
     "chokidar": "^3.5.2",
-    "cssnano": "^5.0.11",
-    "esbuild": "^0.13.14",
-    "eslint": "^8.2.0",
+    "cssnano": "^5.0.12",
+    "esbuild": "^0.14.0",
+    "eslint": "^8.3.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsdoc": "^37.0.3",
@@ -76,15 +76,15 @@
     "material-shadows": "^3.0.1",
     "npm-check-updates": "^12.0.2",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.3.11",
+    "postcss": "^8.4.4",
     "postcss-inline-svg": "^5.0.0",
-    "preact": "^10.5.15",
+    "preact": "^10.6.1",
     "rimraf": "^3.0.2",
-    "sass": "^1.43.4",
+    "sass": "^1.43.5",
     "stylelint": "^14.1.0",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-recommended": "^6.0.0",
-    "stylelint-config-standard-scss": "^2.0.1",
+    "stylelint-config-standard-scss": "^3.0.0",
     "stylelint-scss": "^4.0.0",
     "svgo": "2.3.1",
     "tiny-glob": "^0.2.9",
