@@ -20,8 +20,12 @@
  * IN THE SOFTWARE.
  */
 
-import { Observable, fromEvent } from "rxjs"
-import { map, startWith } from "rxjs/operators"
+import {
+  Observable,
+  fromEvent,
+  map,
+  startWith
+} from "rxjs"
 
 /* ----------------------------------------------------------------------------
  * Types
