@@ -21,20 +21,24 @@
  */
 
 import "focus-visible"
-import { NEVER, Subject, defer, merge } from "rxjs"
 import {
+  EMPTY,
+  NEVER,
+  Subject,
+  defer,
   delay,
   filter,
   map,
+  merge,
   mergeWith,
   shareReplay,
   switchMap
-} from "rxjs/operators"
+} from "rxjs"
 
 import { configuration, feature } from "./_"
 import {
   at,
-  getElement,
+  getOptionalElement,
   requestJSON,
   setToggle,
   watchDocument,
@@ -94,6 +98,7 @@ const keyboard$ = watchKeyboard()
 const viewport$ = watchViewport()
 const tablet$   = watchMedia("(min-width: 960px)")
 const screen$   = watchMedia("(min-width: 1220px)")
+const hover$    = watchMedia("(hover)")
 const print$    = watchPrint()
 
 /* Retrieve search index, if search is enabled */
@@ -137,7 +142,7 @@ keyboard$
         /* Go to previous page */
         case "p":
         case ",":
-          const prev = getElement("[href][rel=prev]")
+          const prev = getOptionalElement("[href][rel=prev]")
           if (typeof prev !== "undefined")
             prev.click()
           break
@@ -145,7 +150,7 @@ keyboard$
         /* Go to next page */
         case "n":
         case ".":
-          const next = getElement("[href][rel=next]")
+          const next = getOptionalElement("[href][rel=next]")
           if (typeof next !== "undefined")
             next.click()
           break
@@ -199,13 +204,13 @@ const content$ = defer(() => merge(
 
   /* Content */
   ...getComponentElements("content")
-    .map(el => mountContent(el, { target$, viewport$, print$ })),
+    .map(el => mountContent(el, { target$, hover$, print$ })),
 
   /* Search highlighting */
   ...getComponentElements("content")
     .map(el => feature("search.highlight")
       ? mountSearchHiglight(el, { index$, location$ })
-      : NEVER
+      : EMPTY
     ),
 
   /* Header title */
@@ -252,8 +257,9 @@ window.location$  = location$          /* Location subject */
 window.target$    = target$            /* Location target observable */
 window.keyboard$  = keyboard$          /* Keyboard observable */
 window.viewport$  = viewport$          /* Viewport observable */
-window.tablet$    = tablet$            /* Tablet observable */
-window.screen$    = screen$            /* Screen observable */
-window.print$     = print$             /* Print mode observable */
+window.tablet$    = tablet$            /* Media tablet observable */
+window.screen$    = screen$            /* Media screen observable */
+window.hover$     = hover$             /* Media hover observable */
+window.print$     = print$             /* Media print observable */
 window.alert$     = alert$             /* Alert subject */
 window.component$ = component$         /* Component observable */
