@@ -21,27 +21,20 @@
  */
 
 import {
-  NEVER,
+  EMPTY,
   Observable,
   Subject,
-  animationFrameScheduler
-} from "rxjs"
-import {
+  defer,
   distinctUntilKeyChanged,
   finalize,
   map,
-  observeOn,
   tap
-} from "rxjs/operators"
+} from "rxjs"
 
 import {
-  resetHeaderTitleState,
-  setHeaderTitleState
-} from "~/actions"
-import {
   Viewport,
-  getElement,
   getElementSize,
+  getOptionalElement,
   watchViewportAt
 } from "~/browser"
 
@@ -94,7 +87,7 @@ interface MountOptions {
 export function watchHeaderTitle(
   el: HTMLHeadingElement, { viewport$, header$ }: WatchOptions
 ): Observable<HeaderTitle> {
-  return watchViewportAt(el, { header$, viewport$ })
+  return watchViewportAt(el, { viewport$, header$ })
     .pipe(
       map(({ offset: { y } }) => {
         const { height } = getElementSize(el)
@@ -120,28 +113,26 @@ export function watchHeaderTitle(
 export function mountHeaderTitle(
   el: HTMLElement, options: MountOptions
 ): Observable<Component<HeaderTitle>> {
-  const internal$ = new Subject<HeaderTitle>()
-  internal$
-    .pipe(
-      observeOn(animationFrameScheduler)
-    )
-      .subscribe(({ active }) => {
-        if (active)
-          setHeaderTitleState(el, "active")
-        else
-          resetHeaderTitleState(el)
-      })
+  return defer(() => {
+    const push$ = new Subject<HeaderTitle>()
+    push$.subscribe(({ active }) => {
+      if (active)
+        el.setAttribute("data-md-state", "active")
+      else
+        el.removeAttribute("data-md-state")
+    })
 
-  /* Obtain headline, if any */
-  const headline = getElement<HTMLHeadingElement>("article h1")
-  if (typeof headline === "undefined")
-    return NEVER
+    /* Obtain headline, if any */
+    const heading = getOptionalElement<HTMLHeadingElement>("article h1")
+    if (typeof heading === "undefined")
+      return EMPTY
 
-  /* Create and return component */
-  return watchHeaderTitle(headline, options)
-    .pipe(
-      tap(state => internal$.next(state)),
-      finalize(() => internal$.complete()),
-      map(state => ({ ref: el, ...state }))
-    )
+    /* Create and return component */
+    return watchHeaderTitle(heading, options)
+      .pipe(
+        tap(state => push$.next(state)),
+        finalize(() => push$.complete()),
+        map(state => ({ ref: el, ...state }))
+      )
+  })
 }
