@@ -20,16 +20,22 @@
  * IN THE SOFTWARE.
  */
 
-import { NEVER, Observable, ObservableInput, merge } from "rxjs"
-import { filter, mergeWith, sample, take } from "rxjs/operators"
+import {
+  NEVER,
+  Observable,
+  ObservableInput,
+  filter,
+  merge,
+  mergeWith,
+  sample,
+  take
+} from "rxjs"
 
 import { configuration } from "~/_"
 import {
   Keyboard,
   getActiveElement,
   getElements,
-  setElementFocus,
-  setElementSelection,
   setToggle
 } from "~/browser"
 import {
@@ -45,10 +51,19 @@ import {
   getComponentElement,
   getComponentElements
 } from "../../_"
-import { SearchQuery, mountSearchQuery } from "../query"
+import {
+  SearchQuery,
+  mountSearchQuery
+} from "../query"
 import { mountSearchResult } from "../result"
-import { SearchShare, mountSearchShare } from "../share"
-import { SearchSuggest, mountSearchSuggest } from "../suggest"
+import {
+  SearchShare,
+  mountSearchShare
+} from "../share"
+import {
+  SearchSuggest,
+  mountSearchSuggest
+} from "../suggest"
 
 /* ----------------------------------------------------------------------------
  * Types
@@ -149,14 +164,14 @@ export function mountSearch(
             case "Escape":
             case "Tab":
               setToggle("search", false)
-              setElementFocus(query, false)
+              query.blur()
               break
 
             /* Vertical arrows: select previous or next search result */
             case "ArrowUp":
             case "ArrowDown":
               if (typeof active === "undefined") {
-                setElementFocus(query)
+                query.focus()
               } else {
                 const els = [query, ...getElements(
                   ":not(details) > [href], summary, details[open] [href]",
@@ -167,7 +182,7 @@ export function mountSearch(
                     key.type === "ArrowUp" ? -1 : +1
                   )
                 ) % els.length)
-                setElementFocus(els[i])
+                els[i].focus()
               }
 
               /* Prevent scrolling of page */
@@ -177,7 +192,7 @@ export function mountSearch(
             /* All other keys: hand to search query */
             default:
               if (query !== getActiveElement())
-                setElementFocus(query)
+                query.focus()
           }
         })
 
@@ -193,8 +208,10 @@ export function mountSearch(
             case "f":
             case "s":
             case "/":
-              setElementFocus(query)
-              setElementSelection(query)
+              query.focus()
+              query.select()
+
+              /* Prevent scrolling of page */
               key.claim()
               break
           }
