@@ -24,7 +24,10 @@ import ClipboardJS from "clipboard"
 import { Observable, Subject } from "rxjs"
 
 import { translation } from "~/_"
-import { getElementOrThrow, getElements } from "~/browser"
+import {
+  getElement,
+  getElements
+} from "~/browser"
 
 /* ----------------------------------------------------------------------------
  * Helper types
@@ -82,7 +85,7 @@ export function setupClipboardJS(
       new ClipboardJS("[data-clipboard-target], [data-clipboard-text]", {
         text: el => (
           el.getAttribute("data-clipboard-text")! ||
-          extract(getElementOrThrow(
+          extract(getElement(
             el.getAttribute("data-clipboard-target")!
           ))
         )
